+++ Problem502/Problem502/ModulusNumber.cs
@@ -8,14 +8,14 @@ namespace Problem502
 {
 	public struct ModulusNumber
 	{
-		public const ulong Mod = 1000000007;
+		public const uint Mod = 1000000007;
 		
 		public static readonly ModulusNumber Zero = new ModulusNumber(0);
 		public static readonly ModulusNumber One = new ModulusNumber(1);
 
-		public ulong Value;
+		public uint Value;
 
-		public ModulusNumber(ulong value)
+		public ModulusNumber(uint value)
 		{
 			Value = value % Mod;
 		}
@@ -47,19 +47,19 @@ namespace Problem502
 
 		public static ModulusNumber operator +(ModulusNumber left, ModulusNumber right)
 		{
-			ulong value = (left.Value + right.Value) % Mod;
+			uint value = (uint)(((ulong)left.Value + (ulong)right.Value) % Mod);
 			return new ModulusNumber(value);
 		}
 
 		public static ModulusNumber operator -(ModulusNumber left, ModulusNumber right)
 		{
-			ulong value = ((Mod + left.Value) - right.Value) % Mod;
+			uint value = (uint)(((Mod + (ulong)left.Value) - (ulong)right.Value) % Mod);
 			return new ModulusNumber(value);
 		}
 
 		public static ModulusNumber operator *(ModulusNumber left, ModulusNumber right)
 		{
-			ulong value = (left.Value * right.Value) % Mod;
+			uint value = (uint)(((ulong)left.Value * (ulong)right.Value) % Mod);
 			return new ModulusNumber(value);
 		}
 	}
